Modify overrides
- Use path definition strings to array
- Move original file only if not symlinks

+++ setup/vim.sh
@@ -7,25 +7,31 @@ gitDir=$HOME/git/dotfiles
 dotVimBundleDir=$HOME/.vim/bundle
 gitDotVimBundleDir=$gitDir/.vim/bundle
 
-# unite-tag
-unite_tag="unite-tag/autoload/unite/sources/tag.vim"
+# ~/.vim/bundle以下の上書き
+overrides=(
+  # unite-tag
+  "unite-tag/autoload/unite/sources/tag.vim"
 
-# phpcomplete-extended
-phpcomplete_extended="phpcomplete-extended/autoload/phpcomplete_extended.vim"
+  # phpcomplete-extended
+  "phpcomplete-extended/autoload/phpcomplete_extended.vim"
+  "phpcomplete-extended/plugin/phpcomplete_extended.vim"
 
-# phpcomplete-extended-symfony
-phpcomplete_extended_symfony="phpcomplete-extended-symfony/autoload/phpcomplete_extended/symfony.vim"
+  # phpcomplete-extended-symfony
+  "phpcomplete-extended-symfony/autoload/phpcomplete_extended/symfony.vim"
 
-# php-getter-setter.vim
-php_getter_setter="php-getter-setter.vim/ftplugin/php_getset.vim"
+  # php-getter-setter.vim
+  "php-getter-setter.vim/ftplugin/php_getset.vim"
 
-# easymotion
-easymotion="vim-easymotion/autoload/EasyMotion.vim"
+  # easymotion
+  "vim-easymotion/autoload/EasyMotion.vim"
+)
 
 # override all
-for override in "$unite_tag" "$phpcomplete_extended" "$phpcomplete_extended_symfony" "$php_getter_setter" "$easymotion"; do
+for override in ${overrides[@]}; do
   if [ -f "${dotVimBundleDir}/${override}" ]; then
-    mv "${dotVimBundleDir}/${override}" "${dotVimBundleDir}/${override}.org"
+    if [ ! -L "${dotVimBundleDir}/${override}" ]; then
+      mv "${dotVimBundleDir}/${override}" "${dotVimBundleDir}/${override}.org"
+    fi
     ln -fs "${gitDotVimBundleDir}/${$override}" "${$dotVimBundleDir}/${override}"
   fi
 done
